# improve coverage for options.key existence