<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

before:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 53.313ms
 PASS  test/key-base64.js 5 OK 220.349ms
 PASS  test/cookie-name.js 6 OK 205.014ms
 PASS  test/key-rotation.js 11 OK 270.789ms
 PASS  test/key.js 10 OK 230.276ms
 PASS  test/access.js 20 OK 265.417ms
 PASS  test/delete.js 14 OK 198.307ms
 PASS  test/dynamic-cookie-options.js 13 OK 258.888ms
 PASS  test/malformed-cookie.js 5 OK 193.168ms
 PASS  test/nonce-bad-length.test.js 5 OK 137.843ms
 PASS  test/salt.js 5 OK 147.088ms
 PASS  test/path.js 4 OK 140.554ms
 PASS  test/schema.js 6 OK 166.027ms
 PASS  test/secret.js 21 OK 3s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  134 passed, of 134
Time:   6s
ERROR: Coverage for lines (95.16%) does not meet global threshold (100%)
ERROR: Coverage for branches (93.75%) does not meet global threshold (100%)
ERROR: Coverage for statements (95.2%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|------------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|------------------------
All files |    95.2 |    93.75 |     100 |   95.16 |
 index.js |    95.2 |    93.75 |     100 |   95.16 | 81,118-119,141-142,262
----------|---------|----------|---------|---------|------------------------
```

after:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 56.457ms
 PASS  test/key-base64.js 5 OK 192.732ms
 PASS  test/key-rotation.js 11 OK 241.602ms
 PASS  test/access.js 20 OK 250.945ms
 PASS  test/dynamic-cookie-options.js 13 OK 210.941ms
 PASS  test/delete.js 14 OK 186.893ms
 PASS  test/cookie-name.js 6 OK 193.142ms
 PASS  test/key.js 12 OK 226.212ms
 PASS  test/malformed-cookie.js 5 OK 169.672ms
 PASS  test/path.js 4 OK 175.181ms
 PASS  test/nonce-bad-length.test.js 5 OK 179.249ms
 PASS  test/schema.js 6 OK 192.387ms
 PASS  test/salt.js 5 OK 178.013ms
 PASS  test/secret.js 21 OK 3s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  136 passed, of 136
Time:   6s
ERROR: Coverage for lines (95.96%) does not meet global threshold (100%)
ERROR: Coverage for branches (95.31%) does not meet global threshold (100%)
ERROR: Coverage for statements (96%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|---------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|---------------------
All files |      96 |    95.31 |     100 |   95.96 |
 index.js |      96 |    95.31 |     100 |   95.96 | 118-119,141-142,262
----------|---------|----------|---------|---------|---------------------
```